Artifacts/app/requirements.txt
shiny>=0.10.2
shinyswatch>=0.7.0
plotly>=5.18.0
pandas>=2.1.4
numpy>=1.26.2
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
fastapi>=0.108.0
uvicorn>=0.25.0
httpx>=0.26.0
gspread>=6.0.0
google-auth>=2.28.0
